refactor(frontend): sub-resource directories
* teams/groups
* users/groups
* groups/events

Each with api, state, and any relevant components.
Hard to review I guess, look at the new structure and if there are
issues with it, and trust the type checker and manual tests and me that
I did it right ^^

Splits some files to prevent them becoming too big.
Will follow up with a recoil refactor based on this restructuring.
