+++ apps/frontend/src/network/groups/state.ts
@@ -5,18 +5,11 @@ import {
   selectorFamily,
   DefaultValue,
 } from 'recoil';
-import {
-  ListGroupResponse,
-  GroupResponse,
-  EventResponse,
-  ListEventResponse,
-} from '@asap-hub/model';
+import { ListGroupResponse, GroupResponse } from '@asap-hub/model';
 
 import { authorizationState } from '@asap-hub/frontend/src/auth/state';
 import { GetListOptions } from '@asap-hub/frontend/src/api-util';
-import { getGroups, getGroup, getGroupEvents } from './api';
-import { BeforeOrAfter } from '../../events/api';
-import { eventState } from '../../events/state';
+import { getGroups, getGroup } from './api';
 
 const groupIndexState = atomFamily<
   { ids: ReadonlyArray<string>; total: number } | Error | undefined,
@@ -59,48 +52,6 @@ export const groupsState = selectorFamily<
     }
   },
 });
-const groupEventIndexState = atomFamily<
-  { ids: ReadonlyArray<string>; total: number } | Error | undefined,
-  GetListOptions & BeforeOrAfter
->({
-  key: 'groupEventIndex',
-  default: undefined,
-});
-
-export const groupEventsState = selectorFamily<
-  ListEventResponse | Error | undefined,
-  GetListOptions & BeforeOrAfter & { id: string }
->({
-  key: 'groupEvents',
-  get: (options) => ({ get }) => {
-    const index = get(groupEventIndexState(options));
-    if (index === undefined || index instanceof Error) return index;
-    const events: EventResponse[] = [];
-    for (const id of index.ids) {
-      const event = get(eventState(id));
-      if (event === undefined) return undefined;
-      events.push(event);
-    }
-    return { total: index.total, items: events };
-  },
-  set: (options) => ({ get, set, reset }, newEvents) => {
-    if (newEvents === undefined || newEvents instanceof DefaultValue) {
-      const oldEvents = get(groupEventIndexState(options));
-      if (!(oldEvents instanceof Error)) {
-        oldEvents?.ids?.forEach((id) => reset(eventState(id)));
-      }
-      reset(groupEventIndexState(options));
-    } else if (newEvents instanceof Error) {
-      set(groupEventIndexState(options), newEvents);
-    } else {
-      newEvents?.items.forEach((event) => set(eventState(event.id), event));
-      set(groupEventIndexState(options), {
-        total: newEvents.total,
-        ids: newEvents.items.map((event) => event.id),
-      });
-    }
-  },
-});
 
 export const refreshGroupState = atomFamily<number, string>({
   key: 'refreshGroup',
@@ -129,25 +80,6 @@ export const useGroups = (options: GetListOptions) => {
     throw groups;
   }
   return groups;
-};
-
-export const useGroupEvents = (
-  id: string,
-  options: GetListOptions & BeforeOrAfter,
-) => {
-  const authorization = useRecoilValue(authorizationState);
-  const [groupEvents, setGroupEvents] = useRecoilState(
-    groupEventsState({ ...options, id }),
-  );
-  if (groupEvents === undefined) {
-    throw getGroupEvents(id, options, authorization)
-      .then(setGroupEvents)
-      .catch(setGroupEvents);
-  }
-  if (groupEvents instanceof Error) {
-    throw groupEvents;
-  }
-  return groupEvents;
 };
 
 export const useGroupById = (id: string) => useRecoilValue(groupState(id));
