+++ apps/frontend/src/network/teams/About.tsx
@@ -1,24 +1,12 @@
 import React from 'react';
 import { join } from 'path';
-import { TeamProfileAbout, TeamGroupsCard } from '@asap-hub/react-components';
+import { TeamProfileAbout } from '@asap-hub/react-components';
 import { TeamResponse } from '@asap-hub/model';
 
 import { SHARED_RESEARCH_PATH, NETWORK_PATH } from '../../routes';
-import { USERS_PATH, GROUPS_PATH } from '../routes';
-import { useTeamGroupsById } from './state';
+import { USERS_PATH } from '../routes';
 import Frame from '../../structure/Frame';
-
-const TeamGroups: React.FC<{ id: string }> = ({ id }) => {
-  const { items, total } = useTeamGroupsById(id);
-  return total > 0 ? (
-    <TeamGroupsCard
-      groups={items.map((group) => ({
-        ...group,
-        href: join('/', NETWORK_PATH, GROUPS_PATH, group.id),
-      }))}
-    />
-  ) : null;
-};
+import GroupsCard from './groups/GroupsCard';
 
 interface AboutProps {
   readonly team: TeamResponse;
@@ -35,7 +23,7 @@ const About: React.FC<AboutProps> = ({ team }) => (
     }
     teamGroupsCard={
       <Frame fallback={null}>
-        <TeamGroups id={team.id} />
+        <GroupsCard id={team.id} />
       </Frame>
     }
   />
