+++ apps/frontend/src/network/teams/__mocks__/api.ts
@@ -1,0 +1,19 @@
+import { TeamResponse, TeamPatchRequest } from '@asap-hub/model';
+import { createTeamResponse } from '@asap-hub/fixtures';
+
+export const getTeam = jest.fn(
+  async (id: string): Promise<TeamResponse> => ({
+    ...createTeamResponse(),
+    id,
+  }),
+);
+
+export const patchTeam = jest.fn(
+  async (id: string, patch: TeamPatchRequest): Promise<TeamResponse> => {
+    const user = await getTeam(id);
+    return {
+      ...user,
+      ...patch,
+    };
+  },
+);
