+++ apps/frontend/src/network/teams/api.ts
@@ -1,9 +1,4 @@
-import {
-  TeamResponse,
-  TeamPatchRequest,
-  ListGroupResponse,
-} from '@asap-hub/model';
-
+import { TeamResponse, TeamPatchRequest } from '@asap-hub/model';
 import { API_BASE_URL } from '../../config';
 
 export const getTeam = async (
@@ -36,21 +31,6 @@ export const patchTeam = async (
   if (!resp.ok) {
     throw new Error(
       `Failed to update team with id ${id}. Expected status 2xx. Received status ${`${resp.status} ${resp.statusText}`.trim()}.`,
-    );
-  }
-  return resp.json();
-};
-
-export const getTeamGroups = async (
-  id: string,
-  authorization: string,
-): Promise<ListGroupResponse> => {
-  const resp = await fetch(`${API_BASE_URL}/teams/${id}/groups`, {
-    headers: { authorization },
-  });
-  if (!resp.ok) {
-    throw new Error(
-      `Failed to fetch team groups with id ${id}. Expected status 2xx. Received status ${`${resp.status} ${resp.statusText}`.trim()}.`,
     );
   }
   return resp.json();
