+++ apps/frontend/src/network/teams/groups/__mocks__/api.ts
@@ -1,0 +1,8 @@
+import { ListGroupResponse } from '@asap-hub/model';
+import { createListGroupResponse } from '@asap-hub/fixtures';
+
+export const getTeamGroups = jest.fn(
+  async (): Promise<ListGroupResponse> => ({
+    ...createListGroupResponse(2),
+  }),
+);
