+++ apps/frontend/src/network/teams/state.ts
@@ -4,13 +4,9 @@ import {
   useRecoilValue,
   useSetRecoilState,
 } from 'recoil';
-import {
-  TeamResponse,
-  TeamPatchRequest,
-  ListGroupResponse,
-} from '@asap-hub/model';
+import { TeamResponse, TeamPatchRequest } from '@asap-hub/model';
 
-import { getTeam, patchTeam, getTeamGroups } from './api';
+import { getTeam, patchTeam } from './api';
 import { authorizationState } from '../../auth/state';
 
 export const refreshTeamState = atomFamily<number, string>({
@@ -37,15 +33,6 @@ const teamState = selectorFamily<TeamResponse | undefined, string>({
     get(patchedTeamState(id)) ?? get(initialTeamState(id)),
 });
 
-const teamGroupsState = selectorFamily<ListGroupResponse, string>({
-  key: 'teamGroups',
-  get: (id) => async ({ get }) => {
-    get(refreshTeamState(id));
-    const authorization = get(authorizationState);
-    return getTeamGroups(id, authorization);
-  },
-});
-
 export const useTeamById = (id: string) => useRecoilValue(teamState(id));
 export const usePatchTeamById = (id: string) => {
   const authorization = useRecoilValue(authorizationState);
@@ -54,5 +41,3 @@ export const usePatchTeamById = (id: string) => {
     setPatchedTeam(await patchTeam(id, patch, authorization));
   };
 };
-export const useTeamGroupsById = (id: string) =>
-  useRecoilValue(teamGroupsState(id));
