+++ apps/frontend/src/network/users/__mocks__/api.ts
@@ -1,12 +1,8 @@
-import {
-  createUserResponse,
-  createListGroupResponse,
-} from '@asap-hub/fixtures';
+import { createUserResponse } from '@asap-hub/fixtures';
 import {
   UserPatchRequest,
   UserResponse,
   UserAvatarPostRequest,
-  ListGroupResponse,
 } from '@asap-hub/model';
 
 export const getUser = jest.fn(
@@ -34,8 +30,4 @@ export const postUserAvatar = jest.fn(
       avatarUrl: `url: ${post.avatar}`,
     };
   },
-);
-
-export const getUserGroups = jest.fn(
-  async (id: string): Promise<ListGroupResponse> => createListGroupResponse(1),
 );
