+++ apps/frontend/src/network/users/api.ts
@@ -2,7 +2,6 @@ import {
   UserResponse,
   UserPatchRequest,
   UserAvatarPostRequest,
-  ListGroupResponse,
 } from '@asap-hub/model';
 
 import { API_BASE_URL } from '../../config';
@@ -62,21 +61,6 @@ export const postUserAvatar = async (
   if (!resp.ok) {
     throw new Error(
       `Failed to update avatar for user with id ${id}. Expected status 2xx. Received status ${`${resp.status} ${resp.statusText}`.trim()}.`,
-    );
-  }
-  return resp.json();
-};
-
-export const getUserGroups = async (
-  id: string,
-  authorization: string,
-): Promise<ListGroupResponse> => {
-  const resp = await fetch(`${API_BASE_URL}/users/${id}/groups`, {
-    headers: { authorization },
-  });
-  if (!resp.ok) {
-    throw new Error(
-      `Failed to fetch groups for user with id ${id}. Expected status 2xx. Received status ${`${resp.status} ${resp.statusText}`.trim()}.`,
     );
   }
   return resp.json();
