+++ apps/frontend/src/network/users/groups/__mocks__/api.ts
@@ -1,0 +1,6 @@
+import { ListGroupResponse } from '@asap-hub/model';
+import { createListGroupResponse } from '@asap-hub/fixtures';
+
+export const getUserGroups = jest.fn(
+  async (id: string): Promise<ListGroupResponse> => createListGroupResponse(1),
+);
