+++ apps/frontend/src/network/users/groups/__tests__/api.test.ts
@@ -1,0 +1,35 @@
+import nock from 'nock';
+import { createListGroupResponse } from '@asap-hub/fixtures';
+
+import { API_BASE_URL } from '@asap-hub/frontend/src/config';
+import { getUserGroups } from '../api';
+
+jest.mock('../../../../config');
+
+afterEach(() => {
+  nock.cleanAll();
+});
+
+describe('getUserGroups', () => {
+  it('makes an authorized GET request for the user id', async () => {
+    nock(API_BASE_URL, { reqheaders: { authorization: 'Bearer x' } })
+      .get('/users/42/groups')
+      .reply(200, {});
+    await getUserGroups('42', 'Bearer x');
+    expect(nock.isDone()).toBe(true);
+  });
+
+  it('returns a successfully fetched user', async () => {
+    const groups = createListGroupResponse();
+    nock(API_BASE_URL).get('/users/42/groups').reply(200, groups);
+    expect(await getUserGroups('42', '')).toEqual(groups);
+  });
+  it('errors for error status', async () => {
+    nock(API_BASE_URL).get('/users/42/groups').reply(500);
+    await expect(
+      getUserGroups('42', ''),
+    ).rejects.toThrowErrorMatchingInlineSnapshot(
+      `"Failed to fetch groups for user with id 42. Expected status 2xx. Received status 500."`,
+    );
+  });
+});
